requirements.txt
aiocache==0.12.3
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
asyncpg==0.27.0
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.8
cryptography==44.0.0
dataclasses-json==0.6.7
Deprecated==1.2.15
dill==0.3.9
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
faiss-cpu==1.9.0.post1
fastapi==0.115.5
filelock==3.16.1
filetype==1.2.0
flake8==7.1.1
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.3
identify==2.6.3
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.1.147
llama-cloud==0.1.5
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.2
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.2
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.0
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.15
loguru==0.7.0
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
mccabe==0.7.0
mock==5.1.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
openai==1.55.3
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20221105
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeps==2.0.1
pyflakes==3.2.0
pylint==3.3.1
pyparsing==3.2.0
pypdf==5.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==4.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.18
pytz==2024.2
PyYAML==6.0.2
redis==4.5.5
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
sentry-sdk==1.16.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.41.3
stdlib-list==0.11.0
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tomlkit==0.13.2
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.22.0
uvloop==0.21.0
virtualenv==20.28.0
watchfiles==1.0.0
websockets==14.1
wrapt==1.17.0
yarl==1.18.3
